CRS-1849: move original breakdown method into new service (tidy up)
